Backend/requirements.txt
amqp==2.1.4
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
billiard==3.5.0.2
celery==4.0.2
channels==1.1.5
constantly==15.1.0
daphne==1.3.0
display==1.0.0
Django==1.10.6
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-common-helpers==0.9.1
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
django-crontab==0.7.1
django-taggit==0.22.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
fcm-django==0.2.11
gcloud==0.17.0
geopy==1.11.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
hyperlink==17.2.1
incremental==17.5.0
jws==0.1.3
kombu==4.0.2
msgpack-python==0.4.8
nose==1.3.7
oauth2client==3.0.0
olefile==0.44
Pillow==4.0.0
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycryptodome==3.4.3
pyfcm==1.3.1
PyJWT==1.5.0
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2017.2
redis==2.10.5
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.11
sseclient==0.0.18
Twisted==17.5.0
txaio==2.8.0
vine==1.1.3
zope.interface==4.4.2
